+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-22T10:01:36.519Z",
+    "2026-07-22T11:53:07.927Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {